Moved cache logic to new ``pipeline_caching.py``
* Population of ``lookup_table``, ``ontology_lookup_table`` and
  ``classification_lookup_table`` moved to ``pipeline_caching.py``

  * New functions functions ``get_predefined_resources``,
    ``get_config_resources`` and ``get_classification_resources``
    respectively

  * Moved in following functions from ``pipeline_helpers.py``:

    * ``get_resource_dict``

    * ``create_lookup_table_skeleton``

    * ``add_predefined_resources_to_lookup_table``

    * ``get_resource_permutation_terms``

    * ``get_resource_bracketed_permutation_terms``

    * ``add_fetched_ontology_to_lookup_table``

  * Moved in ``add_classification_resources_to_lookup_table`` from
    ``pipeline_classification.py``

* Removed last, stale use of ``pkg_resources`` as opposed to using path
  from ``lexmapr.definitions.ROOT``

* Improved in-code documentation in some places

* Modifies tests and imports accordingly
