+++ lexmapr/pipeline_helpers.py
@@ -1,14 +1,12 @@
-"""Helper functions for lexmapr.pipeline."""
+"""Helper functions for lexmapr.pipeline.run."""
 
 from collections import OrderedDict
-import csv
-from itertools import combinations, permutations
+from itertools import combinations
 
 from dateutil.parser import parse
 import inflection
 from nltk.tokenize.moses import MosesDetokenizer
 from nltk.tokenize import word_tokenize
-from pkg_resources import resource_filename
 
 
 def singularize_token(tkn, lookup_table, micro_status):
@@ -231,301 +229,6 @@ def retainedPhrase(termList):
             returnedSet.append(returnItem)
         returnedSetFinal = list(OrderedDict.fromkeys(returnedSet))
     return returnedSetFinal
-
-
-def get_resource_dict(resource_file_name):
-    """Get dictionary containing data from a ``resources/`` csv file.
-
-    The data is standardized to lowercase.
-
-    :param resource_file_name: Name of file (with extension) in
-        ``resources/``
-    :type resource_file_name: str
-    :return: data in ``resource_file_name``
-    :rtype: dict
-    """
-    # Return value
-    ret = {}
-    # Open file_name
-    with open(resource_filename('lexmapr.resources', resource_file_name)) as fp:
-        # Skip first line
-        next(fp)
-        # Read file_name
-        file_contents = csv.reader(fp, delimiter=",")
-        # Iterate across rows in file_contents
-        for row in file_contents:
-            # Get key
-            key = row[0].strip()
-            try:
-                # Get corresponding value
-                val = row[1].strip()
-            except IndexError:
-                # No corresponding value
-                val = ""
-            # Convert key and val to lowercase
-            key = key.lower()
-            val = val.lower()
-            # Add key-value pair to ret
-            ret[key] = val
-    # Return
-    return ret
-
-
-def create_lookup_table_skeleton():
-    """Generate an empty lookup table.
-
-    This means it has all necessary keys, but the values are empty
-    dictionaries.
-
-    :return: Empty lookup table
-    :rtype: dict
-    """
-    return {"synonyms": {},
-            "abbreviations": {},
-            "non_english_words": {},
-            "spelling_mistakes": {},
-            "processes": {},
-            "collocations": {},
-            "inflection_exceptions": {},
-            "stop_words": {},
-            "suffixes": {},
-            "parents": {},
-            "resource_terms_id_based": {},
-            "resource_terms": {},
-            "resource_permutation_terms": {},
-            "resource_bracketed_permutation_terms": {},
-            "buckets_ifsactop": {},
-            "buckets_lexmapr": {},
-            "ifsac_labels": {},
-            "ifsac_refinement": {},
-            "ifsac_default": {}
-            }
-
-
-def add_predefined_resources_to_lookup_table(lookup_table):
-    """Add elements from ``resources/`` to lookup table.
-
-    :param lookup_table: See create_lookup_table_skeleton for the
-                         expected format of this parameter
-    :type lookup_table: dict
-    :return: Modified ``lookup_table``
-    :rtype: dict
-    """
-    # Synonyms of resource terms
-    lookup_table["synonyms"] = get_resource_dict("SynLex.csv")
-    # Abbreviations of resource terms
-    lookup_table["abbreviations"] = get_resource_dict("AbbLex.csv")
-    # Non-english translations of resource terms
-    lookup_table["non_english_words"] = get_resource_dict("NefLex.csv")
-    # Common misspellings of resource terms
-    lookup_table["spelling_mistakes"] = get_resource_dict("ScorLex.csv")
-    # Terms corresponding to candidate processes
-    lookup_table["processes"] = get_resource_dict("candidateProcesses.csv")
-    # Terms corresponding to wikipedia collocations
-    lookup_table["collocations"] = get_resource_dict("wikipediaCollocations.csv")
-    # Terms excluded from inflection treatment
-    lookup_table["inflection_exceptions"] = get_resource_dict("inflection-exceptions.csv")
-    # Constrained list of stop words considered to be meaningless
-    lookup_table["stop_words"] = get_resource_dict("mining-stopwords.csv")
-    # Suffixes to consider appending to terms when mining ontologies
-    lookup_table["suffixes"] = get_resource_dict("suffixes.csv")
-    # ID-resource combinations
-    lookup_table["resource_terms_id_based"] = get_resource_dict("CombinedResourceTerms.csv")
-    # Swap keys and values in resource_terms_id_based
-    lookup_table["resource_terms"] = {
-        v: k for k, v in lookup_table["resource_terms_id_based"].items()
-    }
-
-    # Will contain permutations of resource terms
-    lookup_table["resource_permutation_terms"] = {}
-    # Will contain permutations of resource terms with brackets
-    lookup_table["resource_bracketed_permutation_terms"] = {}
-    # Iterate across resource_terms
-    for resource_term in lookup_table["resource_terms"]:
-        # ID corresponding to resource_term
-        resource_id = lookup_table["resource_terms"][resource_term]
-        # List of tokens in resource_term
-        resource_tokens = word_tokenize(resource_term)
-        # To limit performance overhead, we ignore resource_terms with
-        # more than 7 tokens, as permutating too many tokens can be
-        # costly. We also ignore NCBI taxon terms, as there are
-        # ~160000 such terms.
-        if len(resource_tokens)<7 and "ncbitaxon" not in resource_id:
-            # Add all bracketed permutations of resource_term to
-            # appropriate dictionary.
-            bracketed_perms = get_resource_bracketed_permutation_terms(resource_term)
-            for bracketed_perm in bracketed_perms:
-                lookup_table["resource_bracketed_permutation_terms"][bracketed_perm] = resource_id
-            # Add all permutations of resource_term to appropriate
-            # dictionary.
-            permutations = get_resource_permutation_terms(resource_term)
-            for permutation in permutations:
-                lookup_table["resource_permutation_terms"][permutation] = resource_id
-    return lookup_table
-
-
-def get_resource_permutation_terms(resource_label):
-    """Get permutations of some term.
-
-    :param resource_label: Name of some resource
-    :type resource_label: str
-    :return: All permutations of resource_label
-    :rtype: list
-    """
-    # List of tuples, where each tuple is a different permutation of
-    # tokens from label
-    permutations_set = list(OrderedDict.fromkeys(permutations(resource_label.split())))
-    # Return value
-    ret = []
-    # Generate a string from each tuple, and add it to ret
-    for permutation_tuple in permutations_set:
-        permutation_string = " ".join(permutation_tuple)
-        ret = ret + [permutation_string]
-
-    return ret
-
-
-def get_resource_bracketed_permutation_terms(resource_label):
-    """Get bracketed permutations of some term.
-
-    Bracketed permutations follows the following definition:
-
-    * If a term has no bracketed content, it returns has no bracketed
-      permutations
-    * If a term has bracketed content, the bracketed permutations are
-      comprised of all permutations of the term with the bracket
-      characters removed
-
-    :param resource_label: Name of some resource
-    :type resource_label: str
-    :return: All bracketed permutations of resource_label
-    :rtype: list
-    """
-    if "(" not in resource_label or ")" not in resource_label:
-        return []
-
-    # Portion of label before brackets
-    unbracketed_component = resource_label.split("(")[0]
-    # Portion of label inside brackets
-    bracketed_component = resource_label.split("(")[1]
-    bracketed_component = bracketed_component.split(")")[0]
-    # Replace any commas in bracketed_component with spaces
-    bracketed_component = bracketed_component.replace(",", " ")
-
-    return get_resource_permutation_terms(bracketed_component + " " + unbracketed_component)
-
-
-def add_fetched_ontology_to_lookup_table(lookup_table, fetched_ontology):
-    """Add terms from fetched_ontology to lookup_table.
-
-    lookup_table can be used to map terms in run. See
-    create_lookup_table_skeleton for the expected format of
-    lookup_table.
-
-    :param lookup_table: See create_lookup_table_skeleton for the
-                         expected format of this parameter
-    :param fetched_ontology: See JSON output of ontofetch.py for the
-                             expected format of this parameter
-    :type lookup_table: dict
-    :type fetched_ontology: dict
-    :return: Modified ``lookup_table``
-    :rtype: dict
-    """
-    # Parse content from fetched_ontology and add it to lookup_table
-    for resource in fetched_ontology["specifications"].values():
-        if "id" in resource and "label" in resource:
-            resource_id = resource["id"]
-            resource_label = resource["label"]
-
-            # Standardize values
-            resource_id = resource_id.replace(":", "_")
-            resource_id = resource_id.lower()
-            resource_label = resource_label.lower()
-
-            lookup_table["resource_terms_id_based"][resource_id] = resource_label
-            lookup_table["resource_terms"][resource_label] = resource_id
-
-            # List of tokens in resource_label
-            resource_tokens = word_tokenize(resource_label)
-            # Add permutations if there are less than seven tokens.
-            # Permutating more tokens than this can lead to performance
-            # issues.
-            if len(resource_tokens) < 7:
-                permutations = get_resource_permutation_terms(resource_label)
-                for permutation in permutations:
-                    lookup_table["resource_permutation_terms"][permutation] = resource_id
-
-                bracketed_permutations = get_resource_bracketed_permutation_terms(resource_label)
-                for permutation in bracketed_permutations:
-                    lookup_table["resource_bracketed_permutation_terms"][permutation] = resource_id
-
-            if "oboInOwl:hasSynonym" in resource:
-                synonyms = resource["oboInOwl:hasSynonym"]
-                for synonym in synonyms:
-                    # Standardize synonym
-                    synonym = synonym.lower()
-
-                    lookup_table["synonyms"][synonym] = resource_label
-
-            if "oboInOwl:hasBroadSynonym" in resource:
-                synonyms = resource["oboInOwl:hasBroadSynonym"]
-                for synonym in synonyms:
-                    # Standardize synonym
-                    synonym = synonym.lower()
-
-                    lookup_table["synonyms"][synonym] = resource_label
-
-            if "oboInOwl:hasNarrowSynonym" in resource:
-                synonyms = resource["oboInOwl:hasNarrowSynonym"]
-                for synonym in synonyms:
-                    # Standardize synonym
-                    synonym = synonym.lower()
-
-                    lookup_table["synonyms"][synonym] = resource_label
-
-            if "oboInOwl:hasExactSynonym" in resource:
-                synonyms = resource["oboInOwl:hasExactSynonym"]
-                for synonym in synonyms:
-                    # Standardize synonym
-                    synonym = synonym.lower()
-
-                    lookup_table["synonyms"][synonym] = resource_label
-
-            if "parent_id" in resource:
-                # Standardize parent_id
-                parent_id = resource["parent_id"].replace(":", "_")
-                parent_id = parent_id.lower()
-
-                # Bug in ``ontofetch.py``--sometimes a resource is
-                # parent to itself. Remove when fixed.
-                if resource_id == parent_id:
-                    continue
-                # Instead of overwriting parents like we do with
-                # synonyms, we will concatenate parents from different
-                # fetches.
-                elif resource_id in lookup_table["parents"]:
-                    # Prevent duplicates
-                    if parent_id not in lookup_table["parents"][resource_id]:
-                        lookup_table["parents"][resource_id] += [parent_id]
-                else:
-                    lookup_table["parents"][resource_id] = [parent_id]
-
-                if "other_parents" in resource:
-                    # Standardize values
-                    other_parents = list(map(lambda x: x.replace(":", "_").lower(),
-                                             resource["other_parents"]))
-
-                    # Prevent duplicates
-                    other_parents = list(filter(
-                        lambda x: x not in lookup_table["parents"][resource_id], other_parents))
-
-                    # Bug in ``ontofetch.py``--sometimes a resource is
-                    # parent to itself. Remove when fixed.
-                    other_parents = list(filter(lambda x: x != resource_id, other_parents))
-
-                    lookup_table["parents"][resource_id] += other_parents
-
-    return lookup_table
 
 
 def merge_lookup_tables(lookup_table_one, lookup_table_two):
